requirements.txt
argcomplete==1.12.3
bagit==1.8.1
bcrypt==3.2.0
benford-py==0.5.0
benfords==1.0.2
CacheControl==0.12.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.4
coloredlogs==15.0.1
cwltool==3.1.20211107152837
cycler==0.11.0
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
greenlet==1.1.2
humanfriendly==10.0
idna==3.3
isodate==0.6.0
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
lockfile==0.12.2
lxml==4.6.4
MarkupSafe==2.1.0
matplotlib==3.5.1
mistune==0.8.4
msgpack==1.0.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prov==1.5.1
psutil==5.8.0
pycparser==2.21
pydot==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
rdflib==6.0.2
requests==2.26.0
ruamel.yaml==0.17.17
schema-salad==8.2.20211116214159
scipy==1.8.0
seaborn==0.11.2
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.3
